cyclades: Get all resources accessible by the user
Update Cyclades API to allow access not only to resources that are owned
by the user, but to all resources that are accessible by the user, e.g.
resources that have been shared to a project that the user is a member.
This is implemented by using the 'for_user()' method of the object manager
of the corresponding DB model.

Also, extend log records of all actions to include the user that
performed it. This is required because from now on, the user that
performs an action to a resource is not necessarily the owner of the
resource.

Signed-off-by: Christos Stavrakakis <stavr.chris@gmail.com>