Don't check if the vm fits when allocating
There are 2 possibilities when calling `vm_fits_in_backend`:
1. It returns `False` for every backend, so we
score all of the backends provided.
2. It returns `True` for some of the backends provided.

In the first case `vm_fits_in_backend` was not needed at all,
because we already score all of the VMs.
In the second case since our scoring algorithm is scoring the
backends according to their free space, the backends that we
wouldn't score would get the worst score, so they wouldn't be
returned.

So by removing `vm_fits_in_backend`, we make the `allocate` function
smaller and easier to understand.