ui: Fix networks tab in VM creation wizard
Due to commit 5648edbb if the admin included a public network without
floating IP support in CYCLADES_FORCED_SERVER_NETWORKS the UI tried to
re-add an existing model in a public_networks collection and it crushed.
Fix this by making the appropriate checks prior to adding the created
network in public_networks collection.

Singed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>
Singed-off-by: Kostas Papadimitriou <kpap@grnet.gr>