cyclades: Support "soft enforcing" vm quotas
When a user is overquota on the cyclades.vm resource, we can normally
address the violation by destroying the overquota VMs. As a warning step,
soft enforce allows the admin to shutdown the VMs rather than directly
destroying them.

Introduce a new argument '--soft-resources'. Remove default value of
'--resources'. Now the admin must always specify the resources to be
enforced.