astakos: Introduce membership overquota fields
Introduce fields on project memberships that can be used by a third-party
plugin to handle quota reclamation.

We record a quota-related state. It has a default value, but can be updated
by the plugin whenever an action takes place that can influence the state of
quotas, e.g. when the project is terminated. The plugin can also record the
timestamp for each transition as well as for the first time a membership is
considered overquota, i.e. a state other than the default is set.