astakos: Break down the `change_email` view
Since the `change_email` view handled two cases,
it seemed appropriate to break it into two views.

The first one is used to request an email change.
So an `EmailChange` instance is created and a verification
mail is sent.

The second one is used to verify the email change
the user requested, via the url sent in its mail.