Wrap select_for_update in a transaction
Fix instances of select_for_update that occurred before a transaction
was initialized by decorating the block that runs the
select_for_update with commit_on_success.

This decorator must be placed at the level of views and management
commands. Remove it from the logic level of cyclades volumes app.