Fix clashing 'api' django applications
Django has changed the way it configures and loads applications in 1.7.
See https://docs.djangoproject.com/en/1.7/ref/applications/

As a side-effect, 'pithos.api' app clashes with cyclades
'synnefo.api' app when found on the same node, because they are both
identified as 'api'. To overcome this, use AppConfig to give a prefixed
name to the 'pithos.api' app.