pithos: Cast results of func.sum to long
Postgresql "upgrades" the return type of func.sum. So when 'func.sum' is
applied to a BigInt field it returns a Numeric type which is mapped to
python's Decimal. Since 'func.sum' is applied to a BigInt field and
python's long can hold an arbitrarily large value, it is safe to cast
the return value to long, to better represent the nature of the value
and also make handling it easier.