webproject: Decode isoformatted strings to datetime
Since the Session Serializer serializes dates to isoformatted strings,
when we decode the object, we should convert those strings back to
datetime objects for the framework to use. We can unconditionally
convert all matching strings to date, since we control the session data
and there should be no other matching string that will be wrongfully
converted to datetime object.