cyclades: Modify cached stats calculation
Directly use 'count()' instead of 'len()' when calculating stats.  When
calling len() on a QuerySet, it forces it to be evaluated first and
return the number of elements the resulting list contains. This can be
problematic on a large list of elements, especially when we just need
the number of elements which can be retrieved more efficiently using the
count() call.