admin: Modify url redirect test
Since django passed to Class-based views, the RedirectView function
returned from ResolveMatcher does not take any arguments regarding the
redirect.  Instead the url argument is passed as a class instantiation
parameter which cannot be accessed from the function. Remove the
relevant url checks for now until a solution is found.