Add Keypairs Endpoint
Implemented an endpoint extension for user keypair managment based on
the OpenStack Compute API. This enables all CRUD operations on top of
the keypair objects for easier manipulation. The current setup is not
affected by the changes and it backward compatible.

The new endpoints are accessible through cyclades API on the
/os-keypairs endpoint. Accepted methods are GET(list all keypairs)
GET<id> get information about a specific keypair, POST(create or
update keypair) and DELETE<id> to delete a key