Add API Endpoint Listing
Make the compute API is now explorable from the root of each service. The
responses are based on the ones that the OpenStack API produces. For
example, the Compute service of cyclades is explorable on

{host}/cyclades/compute/

This endpoint will list all the versions supported by Synnefo(only v2.0
in this example). Further exploring is possible by the supported version
number appended to the original root url.

{host}/cyclades/compute/{version}/

Depending on the service, this endpoint will return information such as
supported media types, links of microservices etc.