cyclades: Add keypair name validation
According to Open Stack's specifications, the name of a keypair should
have at least one non whitespace character and only printable characters.
For the time being Synnefo could create a keypair with any name but the URL
`/os-keypairs/{key_name}` would match only if the `key_name` contained,
letters(A-Z, a-z), numbers(0-9), dashes(-) and/or underscores(_).

With the new changes the name is compliant to Open Stack specifications and
it will prevent a user of creating a keypair with an invalid name by returning
him a 400 Bad Request error. Also the underlying URL now matches the new
format.