services: Add new "SNF:exposeVersion" property
This endpoint boolean property can be used when constructing the public
URL of an endpoint to disclude the API version from the resulting URL.
We are using this in some URLs to comply with OpenStack, were the
network and the image service Public URLs are listed by keystone without
the version postfix. Although the official OpenStack clients always
check to see if the version string is included in the URL, there are
third-party clients like the vagrant-openstack-provider that don't.