cyclades: Fix keypair creation functionality
In the previous versions, updating a key (e.g. creating a key with an
existing name) was possible. Open Stack does not allow such an action.
We have updated the creation functionality so that it throws a conflict
exception when a user tries to add a key with a name that already exists
in the database.

Furthermore we have updated the tests so that they match the expected
functionality