cyclades: Add floating IPs endpoint on Compute API
Although in some latter versions of Compute, floating IPs endpoint is
not supported, there are some clients that utilize it and will fail in
any other case (e.g. Vagrant with Openstack driver). Previously in
synnefo, we were implementing this endpoint but the response was
not compliant to the Openstack specifications.

In that sense, we have slightly changed the api methods on
`api/floatingips.py` in order to be more modular. More specifically,
every api method that renders data, accepts a view parameter which is a
function that will format the data appropriately.