Add option to ignore running ssh-agents
If the running ssh-agents have more stored
identities than the MaxAuthTries setting of
sshd_config (which is 6 by default) then
authentication will fail before trying the
generated password for the newly created vm, as
the limit for the authentication tries is met.

With this option set, fabric configures paramiko
to ignore the running ssh-agents and thus
authentication proceeds correctly.