Client side validation of ssh-key name
The check takes place in three cases
- Renaming an existing ssh-key
- Creating a key manually
- Generating a key

Renaming is a generic operation on a view.  When
renaming an object we need to make sure that
object is a key before validating.

Creating and generating a key work through the
submit_key() function, which now returns false if
validation failed, or the newly created model.

The result of submit_key() is used to decide
whether the generate() function should hide the
form (and thus hide the errors) or keep it visible
so the user can fix any errors when they occur.

The validation function validate_form() will do
all the necessary checks and display appropriate
errors in the UI.