cyclades: Add SNF:key_names on server creation
We have implemented a feature that enables multiple SSH key injection
on the server creation. This works alongside with the Open Stack
feature.
NOTE: SNF:key_names and key_name are mutual exclusive, meaning that only
one of them can be set in each request. If both are set then a 400 error
is raised.
We have also extended the test functionality to cover these changes