cyclades: Fix public key validation on UI
We have implemented some updates in order to support more modern
private/public key algorithms such as ed25519. We also made some cleanup
to the code that validates the keys so that it checks the content bytes
and assumes the key type and its validity.

More specifically the validatePublicKey method will do the following:

base64 decode the key contents
From the key header extract the key type
Validated the type of the key
Return the verified key in a key_type key_content formatted string
Also we removed some unnecessary trailing whitespaces from the
underlying file