cyclades: Relax volume type restriction on attach
Relax volume type restriction when attaching a volume.

Before this commit, Synnefo checked that the volume's type is the same
as the one of the VM's flavor. This is stricter than what ganeti
requires, which is simply for all disks to have the same template.

This commit relaxes Synnefo's restriction to match with the Ganeti's
one. This, among others, allows for using volumes from different ext
storage providers (or even the same one with different configuration)
on the same VM.

As a side-effect, it also allows to delete all ExtStorage provided
detached volumes using the same helper which has a specific Volume
type.