cyclades: Add backend parsing for volume type specs
Following the addition of the VolumeTypeSpecs model, in this commit
we are implementing a mechanism to propagate volume type specs in the
lower levels of the Synnefo architecture.

The Cyclades Backend whenever `create_instance` or `attach_volume` is
called, will go through the specs of the underlying volume type and
append any spec starting with a configurable prefix to the settings that
will be propagated to the Ganeti master none.

Also we have implemented some tests for the Volume type spec
functionality both in the scenario when an instance is created as well
as when a volume is attached to an instance.