cyclades: Add flavor specs
Added support for key value pair specs for each flavor type. These
specs can be later used to map flavors to specific backends.

This feature consists of a new model called FlavorSpecs which will
hold the key-value pair data, as well as a migration script to
reflect these changes to the database.

This change is similar to the VolumeTypeSpecs