cyclades: Allow flavor GET if user has VMs using it
Allow fetching flavor details if the user has VMs using the flavor. This
is needed for the UI, to properly display flavor information and decide
on compatible resize flavors. Allowing this does not disclose any extra
information, since the user already knows the details of the Flavor, can
deduce that the flavor is not public and no projects that the user does
not belong are returned.