astakos: Return actual project memberships
When authenticating a token, the user's projects are also returned with
the response. This commit modifies the query to return only the active
project memberships. A user is active member of a project if the project
is active and the membership state is defined as one of the actually
accepted states.