Fix Compass and remove ruby
In order to compile SASS files, in the previous version we where
the ruby binary, Compass which is no longer maintained. As a
result, the Synnefo Admin App installation via the CI kept failing
because of some broken dependency on the underlying package.

In this patch we are completely removing Compass as well as ruby
from build dependencies and we are replacing them with the
`python-libsass` library which consists of bindings for the libsass
package.