astakos: Fix multiple local auth providers
There are some cases in which users, have multiple local auth providers.
In this revision we implemented 2 migration scripts as well as a model
modification in order to delete the duplicates.

The first migration script will identify the affected users and remove
all but one local providers(based on a fixed criterion). The remnant
provider will be used in future authentication attempts from the
underlying user.

The second migration script makes sure that the identifier field of
the provider will never be `null` in order to prevent multiple entries.