astakos: Fix local auth reactivation
Following the migration changes for the users with multiple local auth
providers, we investigated the issues that caused the problem. The main
cause was the fact that a user would create a new local provider
instead of reusing and old one, when trying to reactivate his/hers local
account.

In order to prevent that, when validating the creation/reactivation of
a local provider, we make sure that we will either create a new
provider, or activate an existing one.