cyclades: Handle post-commit/rollback actions
Wrap atomic transaction in the `atomic_context` decorator that also
handles actions that must be performed if the transaction succeeded (or
failed).

Actions are organized in a context. The default context, SerialContext,
provides a simple API to mark the quota serial to be handled, and
finally accept or reject it.

Function issue_and_accept_commission is adapted to make use of the
context.