ui: Fix flavor validation bug
Function `flavor_is_valid` which is responsible for validating a
flavor in terms of quotas, accepts extra quotas. There is a chance
that the object of these extra quotas will be undefined, thus the
function will fail. In order to prevent that, `extra_quotas` variable
now has a default value.