snf_django: Fix printing non-ascii str
Before printing a message, Django's OutputWrapper performs string
operations between the message and the `ending' argument. The default
ending is of unicode type, causing the message to be converted to
unicode in case it's a str. Conversion assumes ascii encoding, failing
if the message contains non-ascii characters.

To overcome this, we set the default ending to a str in order to avoid
the implicit conversion.