cyclades: Parse hvparams status send by eventd
Use the hvparams of the instance's status, as send by eventd to detect
whether the VM is in rescue mode. If the instance has cdrom_image_path
set and the boot_order is set to cdrom, then the VM is considered in
rescue mode. Otherwise, it is not.

In case the cdrom_image_path does not match the value in cyclades DB, we
try to detect it. If that fails, then it is set to None and the VM is
considered rescued with an unknown rescue image.