cyclades: Fix bug in RAPI client
A synnefo-based modification in _SendRequest() casts the url parameters
(query) of a request into a dictionary.  But Ganeti client requests,
such as AddInstanceTags, used to pass url parameters as a list of
attribute-value tuples, e.g. [('tag', 'mytag'), ('tag', 'mytag2')].
Casting this into a dictionary produces collisions, which results in
losing all tags but one because they have the same key.  With the fix
only one tuple is created in the form of ('tag': tags).